added depers_only script: inner work without relication

+++ src/preparations/cleanup_helpers.py
@@ -1,5 +1,8 @@
 import logging
 import psycopg2
+import typing
+
+from src.utils import db_connector
 
 
 logger = logging.getLogger(__name__)
@@ -10,9 +13,14 @@ def src_cleanup_script_helpers(
     src_table: str,
     proccesed_column: str,
     transfer_table: str,
-    publication: str,
+    id_column: str,
+    publication: typing.Optional[str],
+    with_replication: bool,
     after_except: bool,
 ):
+    if after_except:
+        conn.rollback()
+
     logger.info(
         f"Starting cleanup for source-related script helpers, after exception: {after_except}"
     )
@@ -29,11 +37,20 @@ def src_cleanup_script_helpers(
         )
         logger.info(f"Dropped column '{proccesed_column}' from table '{src_table}'")
 
-        cur.execute(f"DROP PUBLICATION {ifExistsClause} {publication}")
-        logger.info(f"Dropped publication '{publication}'")
+        if with_replication:
+            if publication is None:
+                raise Exception("Publication is None, but mode is with replication")
 
-        cur.execute(f"DROP TABLE {ifExistsClause} {transfer_table} CASCADE;")
-        logger.info(f"Dropped table '{transfer_table}'")
+            cur.execute(f"DROP PUBLICATION {ifExistsClause} {publication}")
+            logger.info(f"Dropped publication '{publication}'")
+
+            cur.execute(f"DROP TABLE {ifExistsClause} {transfer_table} CASCADE;")
+            logger.info(f"Dropped table '{transfer_table}'")
+        else:
+            cur.execute(
+                f"ALTER TABLE {transfer_table} DROP COLUMN {ifExistsClause} {id_column} CASCADE;"
+            )
+            logger.info(f"Dropped column '{id_column}' from table '{transfer_table}'")
 
         logger.info("Successfully completed cleanup for source-related script helpers.")
 
@@ -54,6 +71,9 @@ def dst_cleanup_script_helpers(
     subscription: str,
     after_except: bool,
 ):
+    if after_except:
+        conn.rollback()
+
     logger.info(
         f"Starting cleanup for destination-related script helpers, after exception: {after_except}"
     )
@@ -69,9 +89,7 @@ def dst_cleanup_script_helpers(
         cur.execute(f"DROP INDEX CONCURRENTLY {id_column};")
         logger.info(f"Dropped index for column '{id_column}'")
 
-        cur.execute(
-            f"ALTER TABLE {transfer_table} DROP COLUMN {id_column} CASCADE;"
-        )
+        cur.execute(f"ALTER TABLE {transfer_table} DROP COLUMN {id_column} CASCADE;")
         logger.info(f"Dropped column '{id_column}' from table '{transfer_table}'")
 
         logger.info(
@@ -89,26 +107,37 @@ def dst_cleanup_script_helpers(
 
 def cleanup_script_helpers(
     src_conn: psycopg2.extensions.connection,
-    dst_conn: psycopg2.extensions.connection,
+    dst_conn: typing.Optional[db_connector.MultiClusterConnection],
     src_table: str,
     id_column: str,
     processed_column: str,
     transfer_table: str,
-    publication: str,
-    subscription: str,
+    publication: typing.Optional[str],
+    subscription: typing.Optional[str],
+    with_replication: bool,
     after_except: bool = False,
 ):
-    if after_except:
-        src_conn.rollback()
-        dst_conn.rollback()
-
     src_cleanup_script_helpers(
-        src_conn, src_table, processed_column, transfer_table, publication, after_except
-    )
-    dst_cleanup_script_helpers(
-        dst_conn,
+        src_conn,
+        src_table,
+        processed_column,
         transfer_table,
         id_column,
-        subscription,
+        publication,
+        with_replication,
         after_except,
     )
+
+    if with_replication:
+        if dst_conn is None:
+            raise Exception("Dst conn is None, but mode is with replication")
+        if subscription is None:
+            raise Exception("Subscription is None, but mode is with replication")
+
+        dst_cleanup_script_helpers(
+            dst_conn,
+            transfer_table,
+            id_column,
+            subscription,
+            after_except,
+        )
